requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
dnspython==1.15.0
dnspython3==1.15.0
dukpy==0.2.2
geoip2==3.0.0
idna==2.8
importlib-metadata==1.5.0
IPy==1.0
Jinja2==2.11.1
jupyter-echarts-pypkg==0.1.2
jupyter-pip==0.3.1
lml==0.0.2
macropy3==1.1.0b2
MarkupSafe==1.1.1
maxminddb==1.5.2
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
prettytable==0.7.2
py==1.8.1
pyecharts==1.7.1
pyecharts-jupyter-installer==0.0.3
pyparsing==2.4.6
PyQt5==5.9.2
PyQt5-sip==12.7.1
PyQt5-stubs==5.14.1.0
pyqt5-tools==5.13.0.1.5
pytest==5.3.5
pytest-qt==1.2.2
python-dotenv==0.12.0
python-nmap==0.6.1
requests==2.22.0
simplejson==3.17.0
sip==4.19.8
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.1.0
